# hangsha/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Event Calendar API
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    PageMeta:
      type: object
      required: [page, size, total]
      properties:
        page: { type: integer, minimum: 1, example: 1 }
        size: { type: integer, minimum: 1, example: 20 }
        total: { type: integer, minimum: 0, example: 123 }
...
    GetMeResponse:
      type: object
      required: [id]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        profileImageUrl: { type: string, nullable: true }
...
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          nullable: true
          description: |
            닉네임 변경.
            - 필드 미포함: 변경 없음
            - null: 값 비우기(삭제)
        profileImageUrl:
          type: string
          nullable: true
          description: |
            프로필 이미지 URL 변경.
            - 필드 미포함: 변경 없음
            - null: 값 비우기(삭제)
      additionalProperties: false
      description: |
        내 프로필 PATCH.
        요청 바디는 username/profileImageUrl 중 하나 이상을 포함해야 함.

    # ---------------- Categories ----------------
...
    CategoryDto:
      type: object
      required: [id, groupId, name, sortOrder]
      properties:
        id: { type: integer, format: int64 }
        groupId: { type: integer, format: int64 }
        name: { type: string }
        sortOrder: { type: integer, description: "UI에서 고정 순서로 보여주기 위한 정렬값" }
...
    ReplaceAllInterestCategoriesRequest:
      type: object
      required: [ items ]
      properties:
        items:
          type: array
          minItems: 0
          description: "유저의 관심 카테고리 전체 목록을 이 배열로 교체합니다."
          items:
            type: object
            required: [ categoryId, priority ]
            properties:
              categoryId: { type: integer, format: int64 }
              priority:
                type: integer
                minimum: 1
                description: "1이 가장 높음. 권장: 1..N 연속, 중복 없음"
...
    ListInterestCategoryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [category, priority]
            properties:
              category: { $ref: '#/components/schemas/CategoryDto' }
              priority:
                type: integer
                minimum: 1
                description: "유저 관심 카테고리 우선순위 (1이 가장 높음)"

    # ---------------- Events ----------------
    EventDto:
      type: object
      required: [id, title, applyCount]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        imageUrl: { type: string, nullable: true }
        operationMode: { type: string, nullable: true }

        statusId: { type: integer, format: int64, nullable: true }
        eventTypeId: { type: integer, format: int64, nullable: true }
        orgId: { type: integer, format: int64, nullable: true }

        applyStart: { type: string, format: date-time, nullable: true }
        applyEnd: { type: string, format: date-time, nullable: true }
        eventStart: { type: string, format: date-time, nullable: true }
        eventEnd: { type: string, format: date-time, nullable: true }

        capacity: { type: integer, nullable: true }

        applyCount:
          type: integer
          example: 0
          description: |
            현재 이 이벤트에 대해 기록된 신청/관심 수.
            향후 정렬/인기 표시/마감 판단 등에 활용 가능.

        organization: { type: string, nullable: true }
        location: { type: string, nullable: true }
        applyLink: { type: string, nullable: true }

        # --- personalization (Authorization 헤더가 있을 때만 의미) ---
        isInterested:
          type: boolean
          nullable: true
          description: "유저 관심 카테고리와 매칭 여부(토큰 없으면 null 가능)"

        matchedInterestPriority:
          type: integer
          nullable: true
          description: |
            매칭된 관심 카테고리 중 가장 높은 우선순위 값 (1이 가장 높음).
            토큰 없으면 null 가능.

        isBookmarked:
          type: boolean
          nullable: true
          description: "유저가 북마크한 이벤트인지 여부(토큰 없으면 null 가능)"

        tags:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/TagDto' }
          description: |
            (개인화) 유저가 이 이벤트에 대해 추가한 태그 목록.
            - Authorization(로그인)일 때만 의미 있음
            - 비로그인 시 null 가능
    MonthEventResponse:
      type: object
      required: [range, byDate]
      properties:
        range:
          type: object
          required: [from, to]
          properties:
            from: { type: string, format: date }
            to: { type: string, format: date }
        byDate:
          type: object
          additionalProperties:
            type: object
            required: [total, preview]
            properties:
              total: { type: integer }
              preview:
                type: array
                items: { $ref: '#/components/schemas/EventDto' }
    DayEventResponse:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [date, items]
          properties:
            date: { type: string, format: date }
            items:
              type: array
              items: { $ref: '#/components/schemas/EventDto' }
    TitleSearchEventResponse:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/EventDto' }
    DetailEventResponse:
      allOf:
        - $ref: '#/components/schemas/EventDto'
        - type: object
          properties:
            detail:
              type: string
              nullable: true
              description: "모집 공고 상세 글(텍스트)"
...
    TagDto:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: 내 정보 조회
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GetMeResponse' }

    patch:
      tags: [ Users ]
      summary: 내 프로필 수정 (username / profileImageUrl)
      description: |
        - 필드 미포함: 변경 없음
        - null: 해당 값 비우기(삭제)
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
            examples:
              setUsername:
                value: { username: "dyang" }
              clearUsername:
                value: { username: null }
              clearProfileImage:
                value: { profileImageUrl: null }
      responses:
        "204": { description: No Content }
        "400":
          description: Bad Request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ---------------- Categories ----------------
...
  /users/me/interest-categories:
    get:
      tags: [Preference]
      summary: 관심 카테고리 목록 (우선순위 포함)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListInterestCategoryResponse' }

    put:
      tags: [ Preference ]
      summary: 관심 카테고리 전체 교체(배치 저장)
      description: |
        유저의 관심 카테고리 구성을 요청 바디의 items로 "전체 교체"합니다.
        - items가 빈 배열이면 관심 카테고리를 모두 삭제합니다.
        - 권장 규칙: priority는 1..N 연속 + 중복 없음 (서버에서 검증)
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceAllInterestCategoriesRequest' }
      responses:
        "204": { description: No Content }
        "400":
          description: Bad Request (중복 categoryId/priority, priority 규칙 위반 등)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/me/interest-categories/{categoryId}:
    delete:
      tags: [Preference]
      summary: 관심 카테고리 삭제
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: No Content }

  # ---------------- Events ----------------
  /events/month:
    get:
      tags: [Events]
      summary: 월별 캘린더 조회 (Public + Auto personalization on login)
      description: |
        Public API입니다.
        - (로그인 X) 기간(from~to) + 필터(선택) 조건으로 이벤트를 반환합니다.
        - (로그인 O, Authorization: Bearer <token> 제공 시) 개인화가 자동 적용됩니다:
          1) excluded keyword: title에 제외 키워드가 포함된 이벤트는 항상 제외
          2) interest category: 유저 관심 카테고리 priority 기반으로 매칭 이벤트를 항상 상단에 우선 노출
        - 이벤트 preview에는 tags, isBookmarked 등이 포함될 수 있습니다. (토큰 유무/데이터 존재에 따라 null 가능)
      security: []  # Authorization 헤더는 선택(있으면 서버가 personalization 적용)
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, format: date }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date }

        # filtering (optional) - 프로그램 유형/모집현황/주체기관 
        # 복수 카테고리 필터링: &statusId=3&statusId=4 와 같은 방식으로 동작
        - in: query
          name: statusId
          required: false
          schema:
            type: array
            items: { type: integer, format: int64 }
          style: form
          explode: true
        - in: query
          name: eventTypeId
          required: false
          schema:
            type: array
            items: { type: integer, format: int64 }
          style: form
          explode: true
        - in: query
          name: orgId
          required: false
          schema:
            type: array
            items: { type: integer, format: int64 }
          style: form
          explode: true

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthEventResponse' }
  /events/day:
    get:
      tags: [Events]
      summary: 일별 이벤트 조회(페이지네이션) (Public + Auto personalization on login)
      description: |
        Public API입니다.
        - (로그인 X) date + 필터(선택) 조건으로 이벤트를 반환합니다.
        - (로그인 O, Authorization: Bearer <token> 제공 시) 개인화가 자동 적용됩니다:
          1) excluded keyword: title에 제외 키워드가 포함된 이벤트는 항상 제외
          2) interest category: 유저 관심 카테고리 priority 기반으로 매칭 이벤트를 항상 상단에 우선 노출
        - 이벤트 preview에는 tags, isBookmarked 등이 포함될 수 있습니다. (토큰 유무/데이터 존재에 따라 null 가능)
      security: []
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }

        # filtering (optional) - 프로그램 유형/모집현황/주체기관 
        # 복수 카테고리 필터링: &statusId=3&statusId=4 와 같은 방식으로 동작
        - in: query
          name: statusId
          required: false
          schema:
            type: array
            items: { type: integer, format: int64 }
          style: form
          explode: true
        - in: query
          name: eventTypeId
          required: false
          schema:
            type: array
            items: { type: integer, format: int64 }
          style: form
          explode: true
        - in: query
          name: orgId
          required: false
          schema:
            type: array
            items: { type: integer, format: int64 }
          style: form
          explode: true

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DayEventResponse' }
  /events/search/title:
    get:
      tags: [Events]
      summary: 이벤트 제목 검색 (필터링과 별개 기능, Public + Auto personalization on login)
      description: |
        "제목 검색" 전용 API입니다. (필터링 기능과 별개)
        - query는 title contains 검색에 사용됩니다.
        - (로그인 O, Authorization: Bearer <token> 제공 시) 개인화가 자동 적용됩니다:
          1) excluded keyword: title에 제외 키워드가 포함된 이벤트는 항상 제외
          2) interest category: 유저 관심 카테고리 priority 기반으로 매칭 이벤트를 항상 상단에 우선 노출
      security: []
      parameters:
        - in: query
          name: query
          required: true
          schema: { type: string, minLength: 1 }
          description: "이벤트 제목(title) contains 검색"

        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TitleSearchEventResponse' }
  /events/{eventId}:
    get:
      tags: [Events]
      summary: 이벤트 상세 조회 (Public + Optional bookmark state)
      description: |
        Public API입니다.
        - (로그인 X) 이벤트 상세 정보만 반환합니다.
        - (로그인 O + 토큰 제공 시) isBookmarked, bookmarkCount 등 사용자 관련 필드가 채워집니다.
        ※ preference(interest/excluded) 반영 여부는 "목록/검색"에서 의미가 크고, 상세는 주로 bookmark 상태만 optional로 채우는 용도입니다.
      security: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DetailEventResponse' }
        "404":
          description: Not Found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ---------------- Bookmarks ----------------
...